px4_24xxxx_mtd.c: Fix syslog print formatting
ssize_t is either int or long int. Remove compiler warning by casting result
always to int

+++ platforms/nuttx/src/px4/common/px4_24xxxx_mtd.c
@@ -270,7 +270,7 @@ void at24c_test(void)
 			}
 
 		} else if (result != 1) {
-			syslog(LOG_INFO, "unexpected %u\n", result);
+			syslog(LOG_INFO, "unexpected %d\n", (int)result);
 		}
 
 		if ((count % 100) == 0) {
